Upgrades pt. 4
Created scriptable objects for all upgrades in the game, upgrade buttons now update when you buy one

+++ Assets/Scripts/Upgrades/Upgrades.cs
@@ -11,6 +11,6 @@ public class Upgrades : ScriptableObject
     public int cost;
 
     [Header("Multipliers")]
-    public int timeMultiplier;
-    public int clickMultiplier;
+    public float timeMultiplier;
+    public float clickMultiplier;
 }
